+++ app/controllers/users/sessions_controller.rb
@@ -1,7 +1,8 @@
 # frozen_string_literal: true
 
 class Users::SessionsController < ApplicationController
-  before_action: :logged_in?, only: %i(new)
+
+  before_action :session_exists_filter, only: %i(new)
 
   def new
   end
@@ -29,7 +30,16 @@ class Users::SessionsController < ApplicationController
       )
     end
 
-    def logged_in?
+    def session_exists_filter
       redirect_to current_user if logged_in_by_user?
+    end
+
+    def login_by_user(user)
+      session[:user_id] = user.id
+    end
+
+    def logout_user
+      session.delete(:user_id)
+      @current_user = nil
     end
 end
